Regional Language TTS Fine-tuning Project/requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
coqui-tts>=0.6.1
numpy>=1.19.5
pandas>=1.3.0
librosa>=0.8.1
soundfile>=0.10.3
tensorboard>=2.6.0
matplotlib>=3.4.3
scikit-learn>=0.24.2
tqdm>=4.62.2